backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oncoschool-tasks"
version = "0.1.0"
description = "Task Manager for Oncoschool team"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "aiogram>=3.4",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings>=2.1",
    "anthropic>=0.40",
    "openai>=1.10",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "apscheduler>=3.10",
    "pyjwt>=2.8",
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]
